# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Construction Service Coordination API – bilateral coordination between clients (AG) and subcontractors (AN)
...
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /an/leistungsanfragen:
    get:
      operationId: listAnLeistungsanfragen
      tags: [an, leistungsanfragen]
      summary: List local AN Leistungsanfrage projections
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [RECEIVED, DETAILS_RETRIEVED, UNDER_REVIEW, RESPONDED, REVISION_REQUIRED, CONFIRMED, CANCELLED, SUPERSEDED]
      responses:
        "200":
          description: Leistungsanfragen received in the authenticated AN context
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnLeistungsanfrageListItem"
        "403":
          description: Only AN organisations may access local projections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /an/leistungsanfragen/{leistungsanfrageId}/details:
    get:
      operationId: getAnLeistungsanfrageDetails
      tags: [an, leistungsanfragen]
      summary: Get a local AN Leistungsanfrage projection and its resource requirements
      parameters:
        - name: leistungsanfrageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Leistungsanfrage details available in the authenticated AN context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnLeistungsanfrageDetails"
        "403":
          description: Only AN organisations may access local projections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Leistungsanfrage was not received in the AN context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: reviewAnLeistungsanfrageDetails
      tags: [an, leistungsanfragen]
      summary: Mark the published Leistungsanfrage details as consciously reviewed
      description: >
        Completes phase 1 for the authenticated AN. Reading the details remains
        side-effect free; this action records the review and advances RECEIVED
        to DETAILS_RETRIEVED.
      parameters:
        - name: leistungsanfrageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Details were reviewed or had already been reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnLeistungsanfrageListItem"
        "403":
          description: Only AN organisations may access local projections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Leistungsanfrage was not received in the AN context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /takt-requests:
    get:
      operationId: listTaktRequests
      tags: [takt-requests]
      summary: GU lists their own TaktRequests (enriched with Takt, project, NU, transport status)
      description: >
        Returns all TaktRequests for the authenticated GU organisation, joined
        with Takt name, project name, NU organisation name, and the transport
        delivery status of the notification message.

        Access: GU users only. NU users pass role=nu instead; hub admins → 403.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaktRequestStatus"
          description: Filter by request status
        - name: taktId
          in: query
          schema:
            type: string
          description: Filter by Takt ID
        - name: nuOrgId
          in: query
          schema:
            type: string
          description: Filter by NU organisation ID
      responses:
        "200":
          description: List of enriched TaktRequests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaktRequestListItem"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Hub admin or unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createTaktRequestWithSnapshot
      tags: [takt-requests]
      summary: GU creates a TaktRequest draft with an immutable Takt snapshot
      description: >
        Creates a TaktRequest in DRAFT status and atomically persists an
        immutable whitelist-scoped snapshot of the current Takt. No message is
        sent; call /takt-requests/{requestId}/send to deliver the notification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaktRequestBody"
      responses:
        "201":
          description: TaktRequest created in DRAFT status with snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaktRequestDraftResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: GU org is not the project owner, or NU is not a registered contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Takt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Takt has invalid time window or version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /takt-requests/{requestId}/details:
    get:
      operationId: getTaktRequestDetails
      tags: [takt-requests]
      summary: Pull the immutable Takt snapshot released for a TaktRequest
      description: >
        Returns the immutable point-in-time Takt snapshot that was created when
        the request was sent. The response NEVER contains the live Takt row —
        only the snapshot data released by the GU.

        Access:
          - The addressed NU organisation: may access when status is DELIVERED,
            DETAILS_RETRIEVED, or UNDER_REVIEW. First access transitions
            DELIVERED → DETAILS_RETRIEVED and records detailsRetrievedAt.
            Subsequent access (DETAILS_RETRIEVED, UNDER_REVIEW) is idempotent.
          - The creating GU organisation: always has read access for
            control/preview purposes. Does NOT trigger any status transition.

        Forbidden for all other callers (other NU, other GU, hub admins,
        unauthenticated users).
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Immutable Takt snapshot for this TaktRequest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaktRequestDetailsResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the addressed NU or the creating GU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: TaktRequest not found, or snapshot not yet available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Request is in a state that does not permit details retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /takt-requests/{requestId}/availability-checks:
    post:
      operationId: runAvailabilityCheck
      tags: [takt-requests]
      summary: NU runs a feasibility check against their own resource calendar
      description: >
        The addressed NU organisation triggers an availability check for the given
        TaktRequest. The service inspects the NU's resources and existing bookings
        against the planned time window from the snapshot.

        On success the TaktRequest transitions to UNDER_REVIEW (if it was
        DETAILS_RETRIEVED). Repeated calls are allowed (UNDER_REVIEW stays).

        Access: addressed NU only. GU, hub-admin → 403.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Availability check completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityCheckResponse"
        "400":
          description: Check cannot be run (e.g. snapshot missing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the addressed NU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: TaktRequest not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Request is not in a state that allows a check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /takt-requests/{requestId}/availability-checks/latest:
    get:
      operationId: getLatestAvailabilityCheck
      tags: [takt-requests]
      summary: NU retrieves the latest feasibility check result
      description: >
        Returns the most recently completed availability check for the given
        TaktRequest. Prefers COMPLETED checks; falls back to any latest check.

        Access: addressed NU only. GU, hub-admin → 403.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest availability check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityCheckResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the addressed NU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: TaktRequest not found, or no check exists yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /leistungsanfragen/{leistungsanfrageId}/details:
    get:
      operationId: getLeistungsanfrageDetails
      tags: [leistungsanfragen]
      summary: Unveränderlichen Leistungssnapshot abrufen (Alias für GET /takt-requests/:id/details)
      parameters:
        - name: leistungsanfrageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unveränderlicher Leistungssnapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaktRequestDetailsResponse"
        "401":
          description: Nicht authentifiziert
        "403":
          description: Zugriff verweigert
        "404":
          description: Leistungsanfrage oder Snapshot nicht gefunden
        "409":
          description: Status lässt keinen Abruf zu
  /leistungsanfragen/{leistungsanfrageId}/availability-checks:
    post:
      operationId: runLeistungsanfrageAvailabilityCheck
      tags: [leistungsanfragen]
      summary: Verfügbarkeitsprüfung ausführen (Alias für POST /takt-requests/:id/availability-checks)
      parameters:
        - name: leistungsanfrageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Verfügbarkeitsprüfung abgeschlossen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityCheckResponse"
        "400":
          description: Prüfung nicht möglich
        "401":
          description: Nicht authentifiziert
        "403":
          description: Nur AN-Organisationen dürfen Verfügbarkeitsprüfungen ausführen
        "404":
          description: Leistungsanfrage nicht gefunden
        "409":
          description: Status lässt keine Prüfung zu
...
components:
  schemas:
...
    AnLeistungsanfrageDetails:
      allOf:
        - $ref: "#/components/schemas/AnLeistungsanfrageListItem"
        - type: object
          required: [schemaVersion, snapshotPayload, resourceRequirements]
          properties:
            schemaVersion: { type: string }
            snapshotPayload: { type: object, additionalProperties: true }
            resourceRequirements:
              type: array
              items: { $ref: "#/components/schemas/AnLeistungsanfrageResourceRequirement" }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TaktRequestStatus:
      type: string
      description: >
        Status of one specific TaktRequest (coordination request from GU to NU).
        Transport states (DELIVERED) must not be confused with business decisions (ACCEPTED).
        Terminal states: ACCEPTED, CANCELLED, EXPIRED, SUPERSEDED.
      enum:
        - DRAFT
        - SENT
        - DELIVERED
        - DETAILS_RETRIEVED
        - UNDER_REVIEW
        - ACCEPTED
        - ALTERNATIVES_PROPOSED
        - REJECTED
        - REVISION_REQUIRED
        - CANCELLED
        - EXPIRED
        - SUPERSEDED
...
    CreateTaktRequestBody:
      type: object
      description: Body for POST /takt-requests — GU creates a TaktRequest draft with snapshot.
      properties:
        taktId:
          type: string
          minLength: 1
          description: ID of the Takt to coordinate
        nuOrgId:
          type: string
          minLength: 1
          description: ID of the NU organisation to address
        responseRequiredBy:
          type: string
          format: date-time
          description: ISO 8601 deadline by which the NU must respond
          nullable: true
        subject:
          type: string
          maxLength: 255
          description: Optional human-readable subject line for the coordination message
        message:
          type: string
          maxLength: 2000
          description: Optional free-text message from GU to NU
      required:
        - taktId
        - nuOrgId
...